+++ creme/compose/__init__.py
@@ -4,18 +4,14 @@ Meta-estimators for building composite models.
 from .blacklist import Blacklister
 from .func import FuncTransformer
 from .pipeline import Pipeline
-from .target_modifier import TargetModifierRegressor
-from .target_modifier import BoxCoxTransformRegressor
 from .union import TransformerUnion
 from .whitelist import Whitelister
 
 
 __all__ = [
     'Blacklister',
-    'BoxCoxTransformRegressor',
     'FuncTransformer',
     'Pipeline',
-    'TargetModifierRegressor',
     'TransformerUnion',
     'Whitelister'
 ]