+++ creme/optim/vanilla_sgd.py
@@ -11,7 +11,6 @@ class VanillaSGD(base.Optimizer):
 
         ::
 
-            >>> from creme import compose
             >>> from creme import linear_model
             >>> from creme import metrics
             >>> from creme import model_selection
@@ -25,11 +24,11 @@ class VanillaSGD(base.Optimizer):
             ...     shuffle=True,
             ...     random_state=42
             ... )
-            >>> optimizer = optim.VanillaSGD()
-            >>> model = compose.Pipeline([
-            ...     ('scale', preprocessing.StandardScaler()),
-            ...     ('learn', linear_model.LogisticRegression(optimizer))
-            ... ])
+            >>> optimizer = optim.VanillaSGD(0.1)
+            >>> model = (
+            ...     preprocessing.StandardScaler() |
+            ...     linear_model.LogisticRegression(optimizer)
+            ... )
             >>> metric = metrics.F1()
 
             >>> model_selection.online_score(X_y, model, metric)
@@ -40,7 +39,7 @@ class VanillaSGD(base.Optimizer):
 
     """
 
-    def __init__(self, lr=0.1):
+    def __init__(self, lr=0.01):
         super().__init__(lr)
 
     def _update_after_pred(self, w, g):