+++ creme/time_series/__init__.py
@@ -1,0 +1,8 @@
+"""
+Meta-estimators that wrap other estimators.
+"""
+from .detrender import Detrender
+from .detrender import GroupDetrender
+
+
+__all__ = ['Detrender', 'GroupDetrender']